requirements.txt
google-generativeai
python-dotenv
google-adk
Flask
Flask-CORS
requests